+++ src/game/Sprite.java
@@ -9,38 +9,38 @@ import structs.Bounds;
 
 public class Sprite extends Rectangle2D.Double
 {
-    protected final float scale;
-    protected float speed;
-    protected float x_mid_offset;
-    protected float y_mid_offset;
-    protected float x_velocity;
-    protected float y_velocity;
-    protected final float width_scaled;
-    protected final float height_scaled;
+    protected final double scale;
+    protected double speed;
+    protected double x_mid_offset;
+    protected double y_mid_offset;
+    protected double x_velocity;
+    protected double y_velocity;
+    protected final double width_scaled;
+    protected final double height_scaled;
     protected Bounds bounds;
-    protected float radius;
+    protected double radius;
     protected boolean to_remove = false;
 
     private GamePanel panel;
     // Time between images
-    private float delay;
-    private float animation = 0.0f;
+    private double delay;
+    private double animation = 0.0f;
     private BufferedImage[] pics;
     private int current_pic = 0;
     private Vector<Sprite> childs = new Vector<Sprite>();
 
-    public Sprite(GamePanel p, BufferedImage[] imgs, float x, float y, float scale, Bounds b, int delay, float speed)
+    public Sprite(GamePanel p, BufferedImage[] imgs, double x, double y, double scale, Bounds b, int delay, double speed)
     {
         this.speed = speed;
         pics = imgs;
         this.x = x;
         this.y = y;
         this.scale = scale;
-        this.delay = (float) delay;
+        this.delay = (double) delay;
         width = pics[0].getWidth();
         height = pics[0].getHeight();
-        width_scaled = (float) width * scale;
-        height_scaled = (float) height * scale;
+        width_scaled = (double) width * scale;
+        height_scaled = (double) height * scale;
         x_mid_offset = width_scaled / 2.0f;
         y_mid_offset = height_scaled / 2.0f;
         radius = Math.max(width_scaled, height_scaled) / 2.0f;
@@ -48,12 +48,12 @@ public class Sprite extends Rectangle2D.Double
         bounds = b.clone();
     }
 
-    public float distance(Sprite to)
+    public double distance(Sprite to)
     {
-        float a = ((float) to.x + to.x_mid_offset) - ((float) x + x_mid_offset);
-        float b = ((float) to.y + to.y_mid_offset) - ((float) y + y_mid_offset);
+        double a = ((double) to.x + to.x_mid_offset) - ((double) x + x_mid_offset);
+        double b = ((double) to.y + to.y_mid_offset) - ((double) y + y_mid_offset);
 
-        return (float) Math.sqrt(a * a + b * b) - radius - to.radius;
+        return (double) Math.sqrt(a * a + b * b) - radius - to.radius;
     }
 
     private void compute_animation()
@@ -134,8 +134,8 @@ public class Sprite extends Rectangle2D.Double
 
     public void move()
     {
-        float x_moved = x_velocity * panel.deltaTime;
-        float y_moved = y_velocity * panel.deltaTime;
+        double x_moved = x_velocity * panel.deltaTime;
+        double y_moved = y_velocity * panel.deltaTime;
 
         x += x_moved;
         y += y_moved;
